hud: carve panels.rs, focus.rs, widgets.rs (pure moves)
Three slices in one commit (single-tree batch; each is independently a
pure region move, verified with purity.py --worktree before commit):
panel surface builders (9 methods), focus-panel drawing (11), widget
primitives (15, plus the two widget tests). The emptied StationInner
impl shell in mod.rs is dropped; slices mount as bare mod lines
(impl-only modules export nothing).
