+++ crates/station-web/src/hud/mod.rs
@@ -30,6 +30,9 @@ pub(crate) struct Hud {
 }
 
 mod stage;
+mod panels;
+mod focus;
+mod widgets;
 
 #[derive(Default)]
 pub(crate) struct HudStyle {
@@ -143,2937 +146,6 @@ impl Hud {
     pub(crate) fn invalidate(&self) {
         self.invalidate_styles();
         self.invalidate_vignette();
-    }
-}
-
-impl StationInner {
-    /// Actionable surface for a system target's focus panel: header pills
-    /// (panel-wide operations) plus scrollable rows whose clicks and pills
-    /// dispatch the dashboard's real handlers. This is where the
-    /// snapshot's per-domain arrays become *operable* pixels.
-    fn system_panel_surface(&self, id: &str) -> PanelSurface {
-        match id {
-            "system:context" => self.context_panel_surface(),
-            "system:managed" => self.managed_panel_surface(),
-            "system:sessions" => self.sessions_panel_surface(),
-            "system:worktrees" => self.worktrees_panel_surface(),
-            "system:changes" => self.changes_panel_surface(),
-            "system:peers" => self.peers_panel_surface(),
-            "system:activity" => self.activity_panel_surface(),
-            "system:controls" => self.controls_panel_surface(),
-            _ => PanelSurface::default(),
-        }
-    }
-
-    fn sessions_panel_surface(&self) -> PanelSurface {
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new("new", C_TEAL_CSS, HitAction::Composer { op: "open-launch" }),
-                HeaderPill::new(
-                    "refresh",
-                    C_BLUE_CSS,
-                    HitAction::SessionAction {
-                        action: "refresh".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "worktrees",
-                    C_OVERLAY1_CSS,
-                    HitAction::Select("system:worktrees".into()),
-                ),
-            ],
-            empty: "no sessions yet — new launches one",
-            ..Default::default()
-        };
-        for target in self.snapshot.sessions.external_targets.iter() {
-            let sid = nonempty(&target.session_id, &target.id);
-            let mut row = PanelRow::new(
-                truncate(&nonempty(&target.label, "external"), 16),
-                format!("{} / {}", target.value, target.detail),
-                tone_color_css(&target.tone),
-            )
-            .click(HitAction::SessionAction {
-                action: "station-log".into(),
-                id: sid.clone(),
-            });
-            if target.can_focus {
-                row = row.pill(
-                    "focus",
-                    C_PEACH_CSS,
-                    HitAction::SessionAction {
-                        action: "focus".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            if target.can_attach {
-                row = row.pill(
-                    "attach",
-                    C_TEAL_CSS,
-                    HitAction::SessionAction {
-                        action: "attach".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            if target.can_stop {
-                row = row.pill(
-                    "stop",
-                    C_RED_CSS,
-                    HitAction::SessionAction {
-                        action: "stop".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            surface.rows.push(row);
-        }
-        for session in self.snapshot.sessions.recent.iter() {
-            let sid = nonempty(&session.session_id, &session.id);
-            let mut row = PanelRow::new(
-                truncate(&session.value, 16),
-                format!("{} / {}", session.label, session.detail),
-                tone_color_css(&session.tone),
-            )
-            .click(HitAction::SessionAction {
-                action: "station-log".into(),
-                id: sid.clone(),
-            });
-            if session.can_focus {
-                row = row.pill(
-                    "focus",
-                    C_PEACH_CSS,
-                    HitAction::SessionAction {
-                        action: "focus".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            if session.can_resume {
-                row = row.pill(
-                    "resume",
-                    C_GREEN_CSS,
-                    HitAction::SessionAction {
-                        action: "resume".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            // `stop` ends the session gracefully AFTER the current turn;
-            // `interrupt` aborts the turn itself. Mid-turn both intents
-            // matter, so offer both pills when both capabilities exist.
-            if session.can_interrupt {
-                row = row.pill(
-                    "halt",
-                    C_YELLOW_CSS,
-                    HitAction::SessionAction {
-                        action: "interrupt".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            if session.can_stop {
-                row = row.pill(
-                    "stop",
-                    C_RED_CSS,
-                    HitAction::SessionAction {
-                        action: "stop".into(),
-                        id: sid.clone(),
-                    },
-                );
-            } else if session.can_fork {
-                // Codex threads: fork when no lifecycle pill needs the slot
-                // (running sessions keep stop; /fork via the composer covers
-                // them).
-                row = row.pill(
-                    "fork",
-                    C_MAUVE_CSS,
-                    HitAction::SessionAction {
-                        action: "fork".into(),
-                        id: sid.clone(),
-                    },
-                );
-            }
-            surface.rows.push(row);
-        }
-        surface
-    }
-
-    fn worktrees_panel_surface(&self) -> PanelSurface {
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new(
-                    "scan",
-                    C_BLUE_CSS,
-                    HitAction::SessionAction {
-                        action: "worktrees-scan".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "sessions",
-                    C_OVERLAY1_CSS,
-                    HitAction::Select("system:sessions".into()),
-                ),
-            ],
-            empty: "no worktrees scanned — scan discovers them",
-            ..Default::default()
-        };
-        for worktree in self.snapshot.sessions.recent_worktrees.iter() {
-            let path = nonempty(&worktree.id, &worktree.value);
-            surface.rows.push(
-                PanelRow::new(
-                    truncate(&worktree.value, 16),
-                    format!("{} / {}", worktree.label, worktree.detail),
-                    tone_color_css(&worktree.tone),
-                )
-                .click(HitAction::SessionAction {
-                    action: "worktree".into(),
-                    id: path.clone(),
-                })
-                .pill(
-                    "copy",
-                    C_BLUE_CSS,
-                    HitAction::SessionAction {
-                        action: "worktree-copy".into(),
-                        id: path,
-                    },
-                ),
-            );
-        }
-        surface
-    }
-
-    fn activity_panel_surface(&self) -> PanelSurface {
-        let activity = &self.snapshot.activity;
-        let level = nonempty(&activity.level_filter, "all");
-        let next_level = match activity.level_filter.as_str() {
-            "" => "error",
-            "error" => "warn",
-            "warn" => "info",
-            _ => "",
-        };
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new_owned(
-                    format!("lvl {}", truncate(&level, 6)),
-                    C_YELLOW_CSS,
-                    HitAction::ActivityAction {
-                        action: format!("level:{next_level}"),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "copy",
-                    C_BLUE_CSS,
-                    HitAction::ActivityAction {
-                        action: "copy-visible".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "clear",
-                    C_RED_CSS,
-                    HitAction::ActivityAction {
-                        action: "clear-log".into(),
-                        id: String::new(),
-                    },
-                ),
-            ],
-            empty: "no retained activity yet",
-            ..Default::default()
-        };
-        for event in self.snapshot.events.iter().rev() {
-            let color = level_color_css(&event.level);
-            let mut row = PanelRow::new(
-                format!(
-                    "{} {}",
-                    truncate(&nonempty(&event.ts, "--"), 8),
-                    truncate(&event.level, 5)
-                ),
-                truncate(&event.msg, 200),
-                color,
-            )
-            .pill(
-                "copy",
-                C_BLUE_CSS,
-                HitAction::ActivityAction {
-                    action: "copy-event".into(),
-                    id: event.id.clone(),
-                },
-            );
-            if !event.session_id.is_empty() {
-                row = row
-                    .click(HitAction::SessionAction {
-                        action: "station-log".into(),
-                        id: event.session_id.clone(),
-                    })
-                    .pill(
-                        "log",
-                        C_TEAL_CSS,
-                        HitAction::SessionAction {
-                            action: "station-log".into(),
-                            id: event.session_id.clone(),
-                        },
-                    );
-            }
-            surface.rows.push(row);
-        }
-        surface
-    }
-
-    fn context_panel_surface(&self) -> PanelSurface {
-        let context = &self.snapshot.context;
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new(
-                    "live",
-                    C_GREEN_CSS,
-                    HitAction::ContextAction {
-                        action: "live".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "replay",
-                    C_BLUE_CSS,
-                    HitAction::ContextAction {
-                        action: "replay".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "prev",
-                    C_OVERLAY1_CSS,
-                    HitAction::ContextAction {
-                        action: "replay-prev".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "next",
-                    C_OVERLAY1_CSS,
-                    HitAction::ContextAction {
-                        action: "replay-next".into(),
-                        id: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "copy",
-                    C_LAVENDER_CSS,
-                    HitAction::ContextAction {
-                        action: "copy-snapshot".into(),
-                        id: String::new(),
-                    },
-                ),
-            ],
-            empty: "no live context snapshot yet",
-            ..Default::default()
-        };
-        if context.replay_count > 0 {
-            surface.rows.push(PanelRow::new(
-                "replay".to_string(),
-                format!(
-                    "{} / {} of {} / {}",
-                    context.replay_mode,
-                    context.replay_index,
-                    context.replay_count,
-                    nonempty(&context.replay_time, "-")
-                ),
-                C_BLUE_CSS,
-            ));
-        }
-        for cat in context.top_categories.iter() {
-            surface.rows.push(PanelRow::new(
-                truncate(&cat.label, 16),
-                format!(
-                    "{} tok / {} / {}",
-                    fmt_compact(cat.value),
-                    cat.count,
-                    cat.detail
-                ),
-                C_BLUE_CSS,
-            ));
-        }
-        for item in context.top_items.iter() {
-            surface.rows.push(
-                PanelRow::new(
-                    truncate(&item.label, 16),
-                    format!("{} / {}", item.value, item.detail),
-                    tone_color_css(&item.tone),
-                )
-                .pill(
-                    "copy",
-                    C_BLUE_CSS,
-                    HitAction::ContextAction {
-                        action: "copy-part".into(),
-                        id: item.id.clone(),
-                    },
-                ),
-            );
-        }
-        surface
-    }
-
-    fn managed_panel_surface(&self) -> PanelSurface {
-        let managed = &self.snapshot.managed;
-        let state = &managed.action_state;
-        let mut header = vec![HeaderPill::new(
-            "seed",
-            C_BLUE_CSS,
-            HitAction::ManagedAction {
-                action: "seed-context".into(),
-                id: String::new(),
-            },
-        )];
-        if state.can_rewind {
-            header.push(HeaderPill::new(
-                "rewind",
-                C_YELLOW_CSS,
-                HitAction::ManagedAction {
-                    action: "dispatch-rewind".into(),
-                    id: String::new(),
-                },
-            ));
-        }
-        if state.can_backout {
-            header.push(HeaderPill::new(
-                "backout",
-                C_RED_CSS,
-                HitAction::ManagedAction {
-                    action: "run-backout".into(),
-                    id: String::new(),
-                },
-            ));
-        }
-        header.push(HeaderPill::new(
-            "status",
-            C_OVERLAY1_CSS,
-            HitAction::ManagedAction {
-                action: "copy-status".into(),
-                id: String::new(),
-            },
-        ));
-        let mut surface = PanelSurface {
-            header,
-            empty: "no managed rewind records yet",
-            ..Default::default()
-        };
-        if !state.readiness.trim().is_empty() || !state.result.trim().is_empty() {
-            surface.rows.push(PanelRow::new(
-                "state".to_string(),
-                format!(
-                    "{}{}",
-                    state.readiness.trim(),
-                    if state.result.trim().is_empty() {
-                        String::new()
-                    } else {
-                        format!(" / {}", state.result.trim())
-                    }
-                ),
-                C_LAVENDER_CSS,
-            ));
-        }
-        for record in managed.recent_records.iter() {
-            let rid = nonempty(&record.id, &record.label);
-            surface.rows.push(
-                PanelRow::new(
-                    truncate(&record.label, 16),
-                    format!("{} / {}", record.value, record.detail),
-                    tone_color_css(&record.tone),
-                )
-                .click(HitAction::ManagedAction {
-                    action: "record-inspect".into(),
-                    id: rid.clone(),
-                })
-                .pill(
-                    "fork",
-                    C_TEAL_CSS,
-                    HitAction::ManagedAction {
-                        action: "record-fork".into(),
-                        id: rid.clone(),
-                    },
-                )
-                .pill(
-                    "restore",
-                    C_YELLOW_CSS,
-                    HitAction::ManagedAction {
-                        action: "record-restore".into(),
-                        id: rid,
-                    },
-                ),
-            );
-        }
-        surface
-    }
-
-    fn changes_panel_surface(&self) -> PanelSurface {
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new(
-                    "refresh",
-                    C_BLUE_CSS,
-                    HitAction::ChangesAction {
-                        action: "refresh".into(),
-                        path: String::new(),
-                    },
-                ),
-                HeaderPill::new(
-                    "copy paths",
-                    C_OVERLAY1_CSS,
-                    HitAction::ChangesAction {
-                        action: "copy-paths".into(),
-                        path: String::new(),
-                    },
-                ),
-            ],
-            empty: "working tree clean",
-            ..Default::default()
-        };
-        for change in self.snapshot.changes.recent.iter() {
-            let path = nonempty(&change.id, &change.value);
-            surface.rows.push(
-                PanelRow::new(
-                    truncate(&change.label, 16),
-                    format!("{} / {}", change.value, change.detail),
-                    tone_color_css(&change.tone),
-                )
-                .click(HitAction::ChangesAction {
-                    action: "station-diff".into(),
-                    path: path.clone(),
-                })
-                .pill(
-                    "diff",
-                    C_TEAL_CSS,
-                    HitAction::ChangesAction {
-                        action: "station-diff".into(),
-                        path: path.clone(),
-                    },
-                )
-                .pill(
-                    "copy",
-                    C_BLUE_CSS,
-                    HitAction::ChangesAction {
-                        action: "copy-diff".into(),
-                        path,
-                    },
-                ),
-            );
-        }
-        surface
-    }
-
-    fn peers_panel_surface(&self) -> PanelSurface {
-        let runway = &self.snapshot.display_runway;
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new(
-                    "refresh",
-                    C_BLUE_CSS,
-                    HitAction::ControlsAction {
-                        action: "peer-refresh".into(),
-                    },
-                ),
-                HeaderPill::new(
-                    "open",
-                    C_PEACH_CSS,
-                    HitAction::ControlsAction {
-                        action: "peer-open-selected".into(),
-                    },
-                ),
-                HeaderPill::new(
-                    "share",
-                    C_GREEN_CSS,
-                    HitAction::ControlsAction {
-                        action: "display-toggle".into(),
-                    },
-                ),
-            ],
-            empty: "no peers or display lanes yet",
-            ..Default::default()
-        };
-        if !runway.peer_status.trim().is_empty() {
-            surface.rows.push(PanelRow::new(
-                "status".to_string(),
-                runway.peer_status.trim().to_string(),
-                C_OVERLAY1_CSS,
-            ));
-        }
-        for lane in runway.lanes.iter() {
-            let tag = match lane.kind.as_str() {
-                "local_stream" => "local",
-                "remote_stream" => "remote",
-                "peer_target" => "target",
-                "operator_target" => "operator",
-                "shared_view" => "shared",
-                other => other,
-            };
-            let default_op = match lane.kind.as_str() {
-                "peer_target" => "select",
-                "operator_target" | "shared_view" => "focus",
-                _ => "open",
-            };
-            let mut row = PanelRow::new(
-                tag.to_string(),
-                format!("{} / {} / {}", lane.title, lane.meta, lane.detail),
-                if lane.selected {
-                    C_BLUE_CSS
-                } else {
-                    C_PEACH_CSS
-                },
-            );
-            if !lane.id.is_empty() {
-                row = row
-                    .click(HitAction::RunwayAction {
-                        action: default_op.into(),
-                        lane_id: lane.id.clone(),
-                    })
-                    .pill(
-                        "open",
-                        C_PEACH_CSS,
-                        HitAction::RunwayAction {
-                            action: "open".into(),
-                            lane_id: lane.id.clone(),
-                        },
-                    )
-                    .pill(
-                        "copy",
-                        C_BLUE_CSS,
-                        HitAction::RunwayAction {
-                            action: "copy".into(),
-                            lane_id: lane.id.clone(),
-                        },
-                    );
-                if !lane.session_id.is_empty() {
-                    row = row.pill(
-                        "log",
-                        C_TEAL_CSS,
-                        HitAction::SessionAction {
-                            action: "station-log".into(),
-                            id: lane.session_id.clone(),
-                        },
-                    );
-                }
-            }
-            surface.rows.push(row);
-        }
-        surface
-    }
-
-    fn controls_panel_surface(&self) -> PanelSurface {
-        let controls = &self.snapshot.controls;
-        let queue = &self.snapshot.attention_queue;
-        let mut surface = PanelSurface {
-            header: vec![
-                HeaderPill::new(
-                    "compose",
-                    C_BLUE_CSS,
-                    HitAction::Composer { op: "open-send" },
-                ),
-                HeaderPill::new(
-                    "launch",
-                    C_TEAL_CSS,
-                    HitAction::Composer { op: "open-launch" },
-                ),
-            ],
-            empty: "",
-            ..Default::default()
-        };
-
-        // Attention queue first: blocked work is what the operator came for.
-        surface.rows.push(PanelRow::new(
-            "attention".to_string(),
-            format!(
-                "{} blocked / {} warn / {} ready",
-                queue.blocked, queue.warn, queue.ready
-            ),
-            if queue.blocked > 0 {
-                C_RED_CSS
-            } else if queue.warn > 0 {
-                C_YELLOW_CSS
-            } else {
-                C_GREEN_CSS
-            },
-        ));
-        for item in queue.items.iter() {
-            let mut row = PanelRow::new(
-                truncate(&item.level, 16),
-                format!("{} / {} / {}", item.title, item.meta, item.detail),
-                attention_level_color_css(&item.level),
-            );
-            // `log:<session>` targets open that session's transcript;
-            // anything else selects the named scene/system node.
-            if let Some(session_id) = item.target.strip_prefix("log:") {
-                row = row.click(HitAction::SessionAction {
-                    action: "station-log".into(),
-                    id: session_id.to_string(),
-                });
-            } else if !item.target.is_empty() {
-                row = row.click(HitAction::Select(item.target.clone()));
-            }
-            surface.rows.push(row);
-        }
-
-        // Autonomy + backend selection as choice pill rows.
-        surface.rows.push(PanelRow::choices(
-            "autonomy",
-            C_LAVENDER_CSS,
-            ["low", "medium", "high", "full"]
-                .into_iter()
-                .map(|level| {
-                    (
-                        level.to_string(),
-                        controls.autonomy == level,
-                        HitAction::ControlsAction {
-                            action: format!("autonomy:{level}"),
-                        },
-                    )
-                })
-                .collect(),
-        ));
-        surface.rows.push(PanelRow::choices(
-            "backend",
-            C_MAUVE_CSS,
-            [
-                ("intendant", "internal"),
-                ("codex", "codex"),
-                ("claude", "claude-code"),
-            ]
-            .into_iter()
-            .map(|(label, id)| {
-                (
-                    label.to_string(),
-                    // The dashboard reports the no-external-agent state as
-                    // "", "none", or "internal" depending on the source.
-                    controls.backend == id
-                        || (id == "internal"
-                            && matches!(controls.backend.as_str(), "" | "none" | "internal")),
-                    HitAction::ControlsAction {
-                        action: format!("backend:{id}"),
-                    },
-                )
-            })
-            .collect(),
-        ));
-        // Codex runtime rows: visible whenever codex is the global backend
-        // OR the launch composer is aimed at codex — configuring a codex
-        // launch (approval/managed mode) must not require flipping the
-        // global backend first.
-        if controls.backend == "codex" || controls.launch_agent == "codex" {
-            surface.rows.push(PanelRow::choices(
-                "approval",
-                C_YELLOW_CSS,
-                ["untrusted", "on-request", "never"]
-                    .into_iter()
-                    .map(|policy| {
-                        (
-                            policy.to_string(),
-                            controls.approval_policy == policy,
-                            HitAction::ControlsAction {
-                                action: format!("codex-approval:{policy}"),
-                            },
-                        )
-                    })
-                    .collect(),
-            ));
-            surface.rows.push(PanelRow::choices(
-                "managed ctx",
-                C_MAUVE_CSS,
-                ["vanilla", "managed"]
-                    .into_iter()
-                    .map(|mode| {
-                        (
-                            mode.to_string(),
-                            controls.managed_context == mode
-                                || (mode == "vanilla" && controls.managed_context.is_empty()),
-                            HitAction::ControlsAction {
-                                action: format!("codex-managed:{mode}"),
-                            },
-                        )
-                    })
-                    .collect(),
-            ));
-            // Managed mode only works with the Intendant-aware Codex
-            // fork: show which binary managed sessions will spawn, and
-            // flag the ambiguous legacy setup (managed selected, no
-            // dedicated fork configured → falls back to `command`).
-            if controls.managed_context == "managed" {
-                if controls.managed_command.trim().is_empty() {
-                    surface.rows.push(PanelRow::new(
-                        "fork".to_string(),
-                        format!(
-                            "no managed fork set — using {} (must be the Intendant-aware fork)",
-                            nonempty(&controls.command, "codex")
-                        ),
-                        C_YELLOW_CSS,
-                    ));
-                } else {
-                    surface.rows.push(PanelRow::new(
-                        "fork".to_string(),
-                        format!("managed sessions spawn {}", controls.managed_command.trim()),
-                        C_GREEN_CSS,
-                    ));
-                }
-            }
-        }
-        // Claude Code runtime rows: same visibility contract as the Codex
-        // block — global backend OR the launch composer aimed at
-        // claude-code. Model pills are the CLI's latest-version aliases
-        // ("default" clears back to the CLI's own default); permission
-        // pills use short labels for the CLI's camelCase modes.
-        if controls.backend == "claude-code" || controls.launch_agent == "claude-code" {
-            let model = controls.claude_model.trim();
-            const MODEL_ALIASES: [&str; 4] = ["fable", "opus", "sonnet", "haiku"];
-            surface.rows.push(PanelRow::choices(
-                "model",
-                C_YELLOW_CSS,
-                std::iter::once((
-                    "default".to_string(),
-                    model.is_empty(),
-                    HitAction::ControlsAction {
-                        action: "claude-model:default".into(),
-                    },
-                ))
-                .chain(MODEL_ALIASES.into_iter().map(|alias| {
-                    (
-                        alias.to_string(),
-                        model.contains(alias),
-                        HitAction::ControlsAction {
-                            action: format!("claude-model:{alias}"),
-                        },
-                    )
-                }))
-                .collect(),
-            ));
-            // A pinned model outside the alias set still shows truthfully.
-            if !model.is_empty() && !MODEL_ALIASES.iter().any(|alias| model.contains(alias)) {
-                surface.rows.push(PanelRow::new(
-                    "model".to_string(),
-                    format!("custom: {model}"),
-                    C_TEAL_CSS,
-                ));
-            }
-            surface.rows.push(PanelRow::choices(
-                "permissions",
-                C_MAUVE_CSS,
-                [
-                    ("default", "default"),
-                    ("edits", "acceptEdits"),
-                    ("plan", "plan"),
-                    ("bypass", "bypassPermissions"),
-                ]
-                .into_iter()
-                .map(|(label, mode)| {
-                    (
-                        label.to_string(),
-                        controls.claude_permission_mode == mode
-                            || (mode == "default" && controls.claude_permission_mode.is_empty()),
-                        HitAction::ControlsAction {
-                            action: format!("claude-permission:{mode}"),
-                        },
-                    )
-                })
-                .collect(),
-            ));
-        }
-
-        // Voice / video / display sharing toggles.
-        surface.rows.push(PanelRow::choices(
-            "av",
-            C_TEAL_CSS,
-            vec![
-                (
-                    if controls.mic_active {
-                        "mic on"
-                    } else {
-                        "mic off"
-                    }
-                    .to_string(),
-                    controls.mic_active,
-                    HitAction::ControlsAction {
-                        action: "voice-toggle".into(),
-                    },
-                ),
-                (
-                    if controls.video_active {
-                        "cam on"
-                    } else {
-                        "cam off"
-                    }
-                    .to_string(),
-                    controls.video_active,
-                    HitAction::ControlsAction {
-                        action: "video-toggle".into(),
-                    },
-                ),
-                (
-                    "make active".to_string(),
-                    controls.active_browser,
-                    HitAction::ControlsAction {
-                        action: "voice-active".into(),
-                    },
-                ),
-            ],
-        ));
-        surface.rows.push(
-            PanelRow::new(
-                "display".to_string(),
-                format!("share: {}", nonempty(&controls.display_access, "off")),
-                C_PEACH_CSS,
-            )
-            .pill(
-                "toggle",
-                C_PEACH_CSS,
-                HitAction::ControlsAction {
-                    action: "display-toggle".into(),
-                },
-            )
-            .pill(
-                "list",
-                C_BLUE_CSS,
-                HitAction::ControlsAction {
-                    action: "display-list".into(),
-                },
-            ),
-        );
-
-        // Browser workspaces.
-        let mut browser_row = PanelRow::new(
-            "browser".to_string(),
-            format!(
-                "{} workspace{} / {}",
-                controls.browser_workspaces,
-                if controls.browser_workspaces == 1 {
-                    ""
-                } else {
-                    "s"
-                },
-                nonempty(&controls.browser_workspace_status, "idle")
-            ),
-            C_BLUE_CSS,
-        );
-        if controls.browser_workspace_can_create {
-            browser_row = browser_row.pill(
-                "create",
-                C_GREEN_CSS,
-                HitAction::ControlsAction {
-                    action: "browser-create".into(),
-                },
-            );
-        }
-        if controls.browser_workspace_can_acquire {
-            browser_row = browser_row.pill(
-                "acquire",
-                C_TEAL_CSS,
-                HitAction::ControlsAction {
-                    action: "browser-acquire".into(),
-                },
-            );
-        }
-        if controls.browser_workspace_can_close {
-            browser_row = browser_row.pill(
-                "close",
-                C_RED_CSS,
-                HitAction::ControlsAction {
-                    action: "browser-close".into(),
-                },
-            );
-        }
-        if !controls.browser_workspace_url.is_empty() {
-            browser_row = browser_row.pill(
-                "copy",
-                C_BLUE_CSS,
-                HitAction::ControlsAction {
-                    action: "browser-copy".into(),
-                },
-            );
-        }
-        surface.rows.push(browser_row);
-
-        // Recordings: live state + per-stream rows from the side cache.
-        surface.rows.push(
-            PanelRow::new(
-                "recording".to_string(),
-                if controls.active_recording.is_empty() {
-                    format!("{} stored", controls.recordings)
-                } else {
-                    format!(
-                        "recording {} / {} stored",
-                        controls.active_recording, controls.recordings
-                    )
-                },
-                if controls.debug_recording {
-                    C_RED_CSS
-                } else {
-                    C_OVERLAY1_CSS
-                },
-            )
-            .pill(
-                if controls.debug_recording {
-                    "stop rec"
-                } else {
-                    "record"
-                },
-                C_RED_CSS,
-                HitAction::ControlsAction {
-                    action: "debug-record".into(),
-                },
-            )
-            .pill(
-                "screen",
-                C_BLUE_CSS,
-                HitAction::ControlsAction {
-                    action: "debug-screen".into(),
-                },
-            ),
-        );
-        for stream in controls.recording_streams.iter() {
-            let name = nonempty(&stream.action_id, &stream.label);
-            surface.rows.push(
-                PanelRow::new(
-                    truncate(&stream.label, 16),
-                    format!("{} / {}", stream.value, stream.detail),
-                    C_LAVENDER_CSS,
-                )
-                .click(HitAction::ControlsAction {
-                    action: format!("recording-open:{name}"),
-                })
-                .pill(
-                    "play",
-                    C_TEAL_CSS,
-                    HitAction::ControlsAction {
-                        action: format!("recording-open:{name}"),
-                    },
-                ),
-            );
-        }
-
-        // Computer use status.
-        surface.rows.push(PanelRow::new(
-            "computer".to_string(),
-            format!(
-                "{} / {} / {}",
-                nonempty(&controls.cu_backend, "cu"),
-                nonempty(&controls.cu_provider, "provider"),
-                nonempty(&controls.cu_validation_state, "unvalidated")
-            ),
-            C_OVERLAY1_CSS,
-        ));
-        surface
-    }
-
-    /// Shared focus-panel chrome: glass body, FOCUS kicker, title, and the
-    /// close pill (with its hit zones). Body content is the caller's.
-    pub(crate) fn focus_panel_frame(
-        &mut self,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-        title: &str,
-        color: &str,
-    ) {
-        self.glass_panel(
-            x,
-            y,
-            w,
-            h,
-            10.0,
-            hex_color(color).unwrap_or(C_BLUE),
-            1.5,
-            1.1,
-        );
-        self.hit_zones
-            .push(HitZone::new(x, y, w, h, HitAction::Noop));
-        self.text("FOCUS", x + 16.0, y + 23.0, 10.0, C_OVERLAY1_CSS, "bold");
-        self.text(
-            &truncate(title, 34),
-            x + 16.0,
-            y + 47.0,
-            14.0,
-            color,
-            "bold",
-        );
-        self.pill_at(
-            x + w - 70.0,
-            y + 13.0,
-            50.0,
-            23.0,
-            "close",
-            C_OVERLAY1_CSS,
-            false,
-        );
-        self.hit_zones.push(HitZone::new(
-            x + w - 70.0,
-            y + 13.0,
-            50.0,
-            23.0,
-            HitAction::ClosePanel,
-        ));
-    }
-
-    /// Scrollable, actionable rows panel: shared frame + header pills +
-    /// uniform-height rows (click zones, per-row pills, choice pills) +
-    /// scrollbar + a footer inspector line echoing the hovered row in
-    /// full. The workhorse behind every system focus panel.
-    #[allow(clippy::too_many_arguments)]
-    fn rows_panel(
-        &mut self,
-        panel_id: &str,
-        title: &str,
-        color: &str,
-        value: &str,
-        detail: &str,
-        surface: PanelSurface,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-    ) {
-        self.focus_panel_frame(x, y, w, h, title, color);
-        self.text(value, x + 16.0, y + 66.0, 10.5, C_TEXT_CSS, "normal");
-        self.text(
-            &truncate(detail, ((w - 30.0) / 5.8) as usize),
-            x + 16.0,
-            y + 84.0,
-            9.0,
-            C_SUBTEXT0_CSS,
-            "normal",
-        );
-
-        // Header pills: panel-wide operations, left to right.
-        let mut px = x + 16.0;
-        let py = y + 96.0;
-        for pill in &surface.header {
-            let pw = pill.label.chars().count() as f32 * 6.1 + 18.0;
-            if px + pw > x + w - 16.0 {
-                break;
-            }
-            self.pill_at(px, py, pw, 22.0, &pill.label, pill.color, pill.active);
-            self.hit_zones
-                .push(HitZone::new(px, py, pw, 22.0, pill.action.clone()));
-            px += pw + 8.0;
-        }
-
-        // Scrollable rows viewport.
-        let y0 = y + 128.0;
-        let y1 = y + h - 26.0;
-        let viewport_h = (y1 - y0).max(40.0);
-        let rows = &surface.rows;
-        let content_h = rows.len() as f32 * PANEL_ROW_H;
-        self.scroll_zones.push(crate::input::ScrollZone {
-            x,
-            y: y0,
-            w,
-            h: viewport_h,
-            panel: panel_id.to_string(),
-            content_h,
-        });
-        let offset = self.scroll_offset(panel_id, content_h, viewport_h);
-        let scrollable = content_h > viewport_h;
-        let right_pad = if scrollable { 26.0 } else { 18.0 };
-
-        if rows.is_empty() {
-            if !surface.empty.is_empty() {
-                self.text(
-                    surface.empty,
-                    x + 16.0,
-                    y0 + 22.0,
-                    10.0,
-                    C_SUBTEXT0_CSS,
-                    "normal",
-                );
-            }
-            return;
-        }
-
-        let ctx = self.hud.ctx.clone();
-        ctx.save();
-        self.rounded_path(x + 2.0, y0, w - 4.0, viewport_h, 6.0);
-        ctx.clip();
-
-        let first = (offset / PANEL_ROW_H).floor().max(0.0) as usize;
-        let mut hovered_row: Option<usize> = None;
-        for (idx, row) in rows.iter().enumerate().skip(first) {
-            let ry = y0 + idx as f32 * PANEL_ROW_H - offset;
-            if ry > y1 {
-                break;
-            }
-            let hovered = self.hover_xy.is_some_and(|(hx, hy)| {
-                hx >= x && hx <= x + w && hy >= ry && hy <= ry + PANEL_ROW_H && hy >= y0 && hy <= y1
-            });
-            if hovered {
-                hovered_row = Some(idx);
-                self.rounded_path(
-                    x + 8.0,
-                    ry + 1.0,
-                    w - 8.0 - right_pad + 4.0,
-                    PANEL_ROW_H - 3.0,
-                    6.0,
-                );
-                self.hud.set_fill("rgba(137,180,250,0.10)");
-                ctx.fill();
-            }
-            // Label column.
-            self.text(
-                &truncate(&row.label, 17),
-                x + 16.0,
-                ry + 19.0,
-                9.0,
-                row.color,
-                "bold",
-            );
-            // Right-aligned pills; the value text yields to them.
-            let mut pill_x = x + w - right_pad;
-            for pill in row.pills.iter().rev() {
-                let pw = pill.label.chars().count() as f32 * 5.6 + 14.0;
-                pill_x -= pw;
-                if pill_x < x + 130.0 {
-                    break;
-                }
-                self.pill_at(pill_x, ry + 4.5, pw, 21.0, &pill.label, pill.color, false);
-                self.hit_zones.push(HitZone::new(
-                    pill_x,
-                    ry + 4.5,
-                    pw,
-                    21.0,
-                    pill.action.clone(),
-                ));
-                pill_x -= 6.0;
-            }
-            if row.choices.is_empty() {
-                let max_chars = (((pill_x - 6.0) - (x + 124.0)) / 5.7).max(6.0) as usize;
-                self.text(
-                    &truncate(&row.value, max_chars),
-                    x + 124.0,
-                    ry + 19.0,
-                    9.5,
-                    C_TEXT_CSS,
-                    "normal",
-                );
-            } else {
-                // Choice pills row (autonomy / backend / toggles).
-                let mut cx = x + 124.0;
-                for (label, selected, action) in &row.choices {
-                    let cw = label.chars().count() as f32 * 5.8 + 16.0;
-                    if cx + cw > pill_x - 6.0 {
-                        break;
-                    }
-                    self.pill_at(
-                        cx,
-                        ry + 4.5,
-                        cw,
-                        21.0,
-                        label,
-                        if *selected { row.color } else { C_OVERLAY1_CSS },
-                        *selected,
-                    );
-                    self.hit_zones
-                        .push(HitZone::new(cx, ry + 4.5, cw, 21.0, action.clone()));
-                    cx += cw + 6.0;
-                }
-            }
-            // Row body click zone (under the pills, which were pushed after
-            // and therefore win hit-testing).
-            if let Some(action) = &row.click {
-                self.hit_zones.push(HitZone::new(
-                    x + 8.0,
-                    ry,
-                    w - 8.0 - right_pad,
-                    PANEL_ROW_H,
-                    action.clone(),
-                ));
-            }
-        }
-        ctx.restore();
-        self.hud.invalidate_styles();
-
-        if scrollable {
-            self.draw_scrollbar(x + w - 14.0, y0, viewport_h, content_h, offset);
-        }
-
-        // Footer inspector: the hovered row in full, since row values
-        // truncate aggressively next to pills.
-        if let Some(row) = hovered_row.and_then(|idx| rows.get(idx)) {
-            self.text(
-                &truncate(
-                    &format!("{} — {}", row.label, row.value),
-                    ((w - 28.0) / 4.9) as usize,
-                ),
-                x + 16.0,
-                y + h - 9.0,
-                8.5,
-                C_SUBTEXT0_CSS,
-                "normal",
-            );
-        }
-    }
-
-    /// Slim scrollbar: rounded track + position thumb.
-    fn draw_scrollbar(&self, x: f32, y: f32, viewport_h: f32, content_h: f32, offset: f32) {
-        self.hud.set_fill("rgba(49,50,68,0.65)");
-        self.rounded_path(x, y + 2.0, 6.0, viewport_h - 4.0, 3.0);
-        self.hud.ctx.fill();
-        let (thumb_h, thumb_off) =
-            crate::input::scrollbar_thumb(viewport_h - 4.0, content_h, viewport_h, offset);
-        self.hud.set_fill("rgba(137,180,250,0.55)");
-        self.rounded_path(x, y + 2.0 + thumb_off, 6.0, thumb_h, 3.0);
-        self.hud.ctx.fill();
-    }
-
-    /// Composer strip rect for the current mode: `(x, y, w, h)`. Shared
-    /// between the strip painter and the transcript panel (which yields
-    /// vertical space when both are open).
-    pub(crate) fn composer_rect(&self, w: f32, h: f32) -> (f32, f32, f32, f32) {
-        let lane_y = (h - lane_metrics(self.density, h).2 - 24.0).max(282.0);
-        let strip_h = if self.composer_mode == "launch" {
-            96.0
-        } else {
-            56.0
-        };
-        let sw = (w * 0.52).clamp(320.0, 660.0);
-        (24.0, lane_y - strip_h - 12.0, sw, strip_h)
-    }
-
-    /// The composer strip: canvas-drawn chrome for the DOM input overlay.
-    /// Send mode: target chip + input slot + send. Launch mode: input slot
-    /// + agent choice pills + execution pills + launch.
-    fn draw_composer_strip(&mut self, w: f32, h: f32) {
-        let (x, y, sw, sh) = self.composer_rect(w, h);
-        let controls = &self.snapshot.controls;
-        let launch = self.composer_mode == "launch";
-        self.glass_panel(x, y, sw, sh, 12.0, C_BLUE, 1.8, 1.08);
-        self.hit_zones
-            .push(HitZone::new(x, y, sw, sh, HitAction::Noop));
-
-        let kicker = if launch {
-            let missing = controls.launch_missing.trim();
-            if controls.launch_ready || missing.is_empty() {
-                "LAUNCH NEW SESSION".to_string()
-            } else {
-                format!("LAUNCH NEW SESSION — needs {}", truncate(missing, 28))
-            }
-        } else {
-            format!("COMPOSE → {}", truncate(&self.station_target_label(), 36))
-        };
-        self.text(&kicker, x + 16.0, y + 16.0, 8.0, C_BLUE_CSS, "bold");
-        self.text(
-            "enter sends / esc closes",
-            x + sw - 150.0,
-            y + 16.0,
-            7.5,
-            C_OVERLAY1_CSS,
-            "normal",
-        );
-
-        // Input slot: dark inset the DOM textarea sits over.
-        let action_w = if launch { 88.0 } else { 76.0 };
-        let slot_x = x + 14.0;
-        let slot_w = sw - 28.0 - action_w - 10.0;
-        let slot_y = y + 24.0;
-        self.rounded_path(slot_x, slot_y, slot_w, 24.0, 7.0);
-        self.hud.set_fill("rgba(9,10,18,0.78)");
-        self.hud.ctx.fill();
-        self.rounded_path(slot_x, slot_y, slot_w, 24.0, 7.0);
-        self.hud.set_stroke("rgba(137,180,250,0.35)");
-        self.hud.ctx.stroke();
-        self.composer_input_rect = Some((slot_x + 6.0, slot_y + 2.0, slot_w - 12.0, 20.0));
-
-        let send_x = slot_x + slot_w + 10.0;
-        if launch {
-            self.pill_at(
-                send_x,
-                slot_y + 1.0,
-                action_w,
-                22.0,
-                "launch",
-                C_TEAL_CSS,
-                true,
-            );
-            self.hit_zones.push(HitZone::new(
-                send_x,
-                slot_y + 1.0,
-                action_w,
-                22.0,
-                HitAction::Composer { op: "launch" },
-            ));
-        } else {
-            let label = if controls.prompt_mode == "steer" {
-                "steer"
-            } else {
-                "send"
-            };
-            self.pill_at(
-                send_x,
-                slot_y + 1.0,
-                action_w,
-                22.0,
-                label,
-                C_BLUE_CSS,
-                true,
-            );
-            self.hit_zones.push(HitZone::new(
-                send_x,
-                slot_y + 1.0,
-                action_w,
-                22.0,
-                HitAction::Composer { op: "send" },
-            ));
-        }
-
-        if launch {
-            // Agent choice pills + execution shape pills. An empty
-            // launch_mode means execution does not apply (external agent
-            // selected); the agent pills then reclaim the full row.
-            self.text("agent", x + 16.0, y + 70.0, 8.0, C_TEAL_CSS, "bold");
-            let execution = controls.launch_mode.as_str();
-            let exec_pills = [
-                ("auto", "auto", C_TEAL_CSS),
-                ("orch", "orchestrate", C_LAVENDER_CSS),
-                ("direct", "direct", C_PEACH_CSS),
-            ];
-            let exec_w = exec_pills
-                .iter()
-                .map(|(label, _, _)| label.chars().count() as f32 * 5.8 + 16.0 + 4.0)
-                .sum::<f32>()
-                - 4.0;
-            let agent_limit = if execution.is_empty() {
-                x + sw - 16.0
-            } else {
-                x + sw - exec_w - 24.0
-            };
-            let mut cx = x + 58.0;
-            let selected_agent = controls.launch_agent.as_str();
-            for (label, id) in [
-                ("auto", ""),
-                ("intendant", "internal"),
-                ("codex", "codex"),
-                ("claude", "claude-code"),
-            ] {
-                let cw = label.chars().count() as f32 * 5.8 + 16.0;
-                if cx + cw > agent_limit {
-                    break;
-                }
-                let active = selected_agent == id;
-                self.pill_at(
-                    cx,
-                    y + 58.0,
-                    cw,
-                    21.0,
-                    label,
-                    if active { C_TEAL_CSS } else { C_OVERLAY1_CSS },
-                    active,
-                );
-                self.hit_zones.push(HitZone::new(
-                    cx,
-                    y + 58.0,
-                    cw,
-                    21.0,
-                    HitAction::ControlsAction {
-                        action: format!("launch-agent:{id}"),
-                    },
-                ));
-                cx += cw + 6.0;
-            }
-            if !execution.is_empty() {
-                let mut ex = x + sw - exec_w - 16.0;
-                for (label, value, accent) in exec_pills {
-                    let cw = label.chars().count() as f32 * 5.8 + 16.0;
-                    let active = execution == value;
-                    self.pill_at(
-                        ex,
-                        y + 58.0,
-                        cw,
-                        21.0,
-                        label,
-                        if active { accent } else { C_OVERLAY1_CSS },
-                        active,
-                    );
-                    self.hit_zones.push(HitZone::new(
-                        ex,
-                        y + 58.0,
-                        cw,
-                        21.0,
-                        HitAction::ControlsAction {
-                            action: format!("launch-execution:{value}"),
-                        },
-                    ));
-                    ex += cw + 4.0;
-                }
-            }
-        } else {
-            // Target chip: click opens the sessions panel to retarget.
-            let chip_w = 70.0;
-            self.pill_at(
-                x + sw - 78.0 - action_w,
-                y + 1.0,
-                chip_w,
-                18.0,
-                "target",
-                C_PEACH_CSS,
-                false,
-            );
-            self.hit_zones.push(HitZone::new(
-                x + sw - 78.0 - action_w,
-                y + 1.0,
-                chip_w,
-                18.0,
-                HitAction::Composer { op: "target" },
-            ));
-        }
-    }
-
-    /// Transcript / diff viewer: a large left-anchored panel with
-    /// word-wrapped, kind-colored rows and pixel scrolling. Content
-    /// layout (wrapping) is cached per (content, width) signature.
-    fn draw_transcript_panel(&mut self, w: f32, h: f32) {
-        let Some(transcript) = self.transcript.clone() else {
-            return;
-        };
-        let lane_y = (h - lane_metrics(self.density, h).2 - 24.0).max(282.0);
-        let x = 24.0;
-        let tw = (w * 0.56).clamp(340.0, 820.0).min(w - 48.0);
-        let top = 58.0 + 14.0;
-        let bottom = if self.composer_open {
-            self.composer_rect(w, h).1 - 10.0
-        } else {
-            lane_y - 12.0
-        };
-        let th = (bottom - top).max(180.0);
-        let diff = transcript.mode == "diff";
-        let accent = if diff { C_YELLOW_CSS } else { C_TEAL_CSS };
-
-        self.glass_panel(
-            x,
-            top,
-            tw,
-            th,
-            12.0,
-            hex_color(accent).unwrap_or(C_TEAL),
-            1.4,
-            1.06,
-        );
-        self.hit_zones
-            .push(HitZone::new(x, top, tw, th, HitAction::Noop));
-        self.text(
-            if diff { "DIFF" } else { "TRANSCRIPT" },
-            x + 16.0,
-            top + 21.0,
-            10.0,
-            C_OVERLAY1_CSS,
-            "bold",
-        );
-        self.text(
-            &truncate(&nonempty(&transcript.label, &transcript.session_id), 42),
-            x + 16.0,
-            top + 43.0,
-            13.0,
-            accent,
-            "bold",
-        );
-        self.pill_at(
-            x + tw - 66.0,
-            top + 12.0,
-            50.0,
-            22.0,
-            "close",
-            C_OVERLAY1_CSS,
-            false,
-        );
-        self.hit_zones.push(HitZone::new(
-            x + tw - 66.0,
-            top + 12.0,
-            50.0,
-            22.0,
-            HitAction::CloseTranscript,
-        ));
-        // Header ops.
-        let mut hx = x + 16.0;
-        let header_pills: Vec<(&str, &str, HitAction)> = if diff {
-            vec![(
-                "copy diff",
-                C_BLUE_CSS,
-                HitAction::ChangesAction {
-                    action: "copy-diff".into(),
-                    path: transcript.session_id.clone(),
-                },
-            )]
-        } else {
-            vec![
-                ("steer", C_BLUE_CSS, HitAction::Composer { op: "open-send" }),
-                (
-                    "focus",
-                    C_PEACH_CSS,
-                    HitAction::SessionAction {
-                        action: "focus".into(),
-                        id: transcript.session_id.clone(),
-                    },
-                ),
-                (
-                    "copy id",
-                    C_OVERLAY1_CSS,
-                    HitAction::SessionAction {
-                        action: "copy".into(),
-                        id: transcript.session_id.clone(),
-                    },
-                ),
-            ]
-        };
-        for (label, color, action) in header_pills {
-            let pw = label.chars().count() as f32 * 6.1 + 18.0;
-            self.pill_at(hx, top + 52.0, pw, 22.0, label, color, false);
-            self.hit_zones
-                .push(HitZone::new(hx, top + 52.0, pw, 22.0, action));
-            hx += pw + 8.0;
-        }
-        self.text(
-            &format!(
-                "{} of {} entries",
-                transcript.rows.len(),
-                transcript.total.max(transcript.rows.len() as u32)
-            ),
-            x + tw - 190.0,
-            top + 66.0,
-            8.5,
-            C_SUBTEXT0_CSS,
-            "normal",
-        );
-
-        let y0 = top + 84.0;
-        let y1 = top + th - 14.0;
-        let viewport_h = (y1 - y0).max(60.0);
-
-        if !transcript.error.is_empty() {
-            self.text(
-                &truncate(&transcript.error, ((tw - 30.0) / 5.8) as usize),
-                x + 16.0,
-                y0 + 18.0,
-                10.0,
-                C_RED_CSS,
-                "normal",
-            );
-            return;
-        }
-        if transcript.rows.is_empty() {
-            self.text(
-                "no entries — waiting for the session log",
-                x + 16.0,
-                y0 + 18.0,
-                10.0,
-                C_SUBTEXT0_CSS,
-                "normal",
-            );
-            return;
-        }
-
-        // Wrapped layout, cached against a cheap content signature.
-        let gutter = if diff { 16.0 } else { 64.0 };
-        let wrap_px = tw - gutter - 16.0 - 18.0;
-        let wrap_chars = ((wrap_px / 5.6).max(20.0)) as usize;
-        let sig = transcript_signature(&transcript, wrap_chars as u32);
-        if self.transcript_layout.as_ref().map(|l| l.sig) != Some(sig) {
-            let mut layout = layout_transcript(&transcript, wrap_chars);
-            layout.sig = sig;
-            self.transcript_layout = Some(layout);
-        }
-        let layout = self.transcript_layout.as_ref().unwrap();
-        let content_h = layout.content_h;
-        self.scroll_zones.push(crate::input::ScrollZone {
-            x,
-            y: y0,
-            w: tw,
-            h: viewport_h,
-            panel: "transcript".to_string(),
-            content_h,
-        });
-        if self.transcript_follow {
-            let max = (content_h - viewport_h).max(0.0);
-            self.panel_scroll.insert("transcript".to_string(), max);
-        }
-        let offset = self.scroll_offset("transcript", content_h, viewport_h);
-
-        let ctx = self.hud.ctx.clone();
-        ctx.save();
-        self.rounded_path(x + 2.0, y0, tw - 4.0, viewport_h, 6.0);
-        ctx.clip();
-        let layout = self.transcript_layout.take().unwrap();
-        for line in &layout.lines {
-            let ly = y0 + line.y - offset;
-            if ly < y0 - TRANSCRIPT_LINE_H {
-                continue;
-            }
-            if ly > y1 + TRANSCRIPT_LINE_H {
-                break;
-            }
-            let color = transcript_kind_color(&line.kind);
-            if line.first && !diff {
-                self.text(&truncate(&line.kind, 9), x + 16.0, ly, 8.0, color, "bold");
-                if !line.ts.is_empty() {
-                    self.text(
-                        &truncate(&line.ts, 8),
-                        x + 16.0,
-                        ly + 9.0,
-                        6.5,
-                        C_OVERLAY1_CSS,
-                        "normal",
-                    );
-                }
-            }
-            self.text(
-                &line.text,
-                x + gutter,
-                ly,
-                9.5,
-                if diff {
-                    color
-                } else if line.kind == "user" {
-                    C_TEXT_CSS
-                } else {
-                    color
-                },
-                if line.first && line.kind == "user" {
-                    "bold"
-                } else {
-                    "normal"
-                },
-            );
-        }
-        self.transcript_layout = Some(layout);
-        ctx.restore();
-        self.hud.invalidate_styles();
-
-        if content_h > viewport_h {
-            self.draw_scrollbar(x + tw - 14.0, y0, viewport_h, content_h, offset);
-        }
-    }
-
-    /// One labeled row inside a focus panel: colored label column, value
-    /// text beside it. Returns the next row baseline.
-    fn focus_row(&self, x: f32, row_y: f32, w: f32, label: &str, value: &str, color: &str) -> f32 {
-        self.text(&truncate(label, 11), x + 16.0, row_y, 9.0, color, "bold");
-        self.text(
-            &truncate(value, ((w - 116.0) / 5.6).max(18.0) as usize),
-            x + 96.0,
-            row_y,
-            9.5,
-            C_TEXT_CSS,
-            "normal",
-        );
-        row_y + 17.0
-    }
-
-    /// Real detail panel for a selected agent node: identity, model, phase,
-    /// task, budget/usage, and — when an approval is pending — the approval
-    /// command plus actionable approve/deny pills.
-    fn draw_agent_focus(
-        &mut self,
-        agent: &crate::model::StationAgent,
-        x: f32,
-        panel_w: f32,
-        activity_lane_y: f32,
-    ) {
-        let approval = agent.needs_approval
-            && (agent.host_id == "local"
-                || self
-                    .snapshot
-                    .hosts
-                    .first()
-                    .is_some_and(|h| h.id == agent.host_id)
-                || agent
-                    .approval_id
-                    .as_deref()
-                    .is_some_and(|id| !id.is_empty()));
-        let is_session = !agent.session_id.trim().is_empty();
-        let has_goal =
-            !agent.goal_objective.trim().is_empty() || !agent.goal_status.trim().is_empty();
-        let has_cache_vitals = agent.cache_hit_pct >= 0.0
-            || (agent.cache_ttl_seconds > 0.0 && agent.cache_last_activity_epoch > 0.0);
-        let rows = 5
-            + usize::from(!agent.worktree.trim().is_empty())
-            + usize::from(has_goal)
-            + usize::from(!agent.relationship_kind.trim().is_empty())
-            + usize::from(!agent.vitals_git.trim().is_empty())
-            + usize::from(has_cache_vitals)
-            + usize::from(!agent.vitals_limits.trim().is_empty())
-            + if approval { 2 } else { 0 };
-        let panel_h = 74.0
-            + rows as f32 * 17.0
-            + if approval { 30.0 } else { 6.0 }
-            + if is_session { 34.0 } else { 0.0 };
-        let y = (activity_lane_y - panel_h - 12.0).max(58.0);
-        let phase = phase_color_css(&agent.phase);
-        self.focus_panel_frame(x, y, panel_w, panel_h, &agent.id, phase);
-        let subtitle = if is_session && agent.recent {
-            format!("recent {} session", nonempty(&agent.source, "intendant"))
-        } else if is_session {
-            format!("{} session", nonempty(&agent.source, "intendant"))
-        } else {
-            format!("{} agent", nonempty(&agent.role, "agent"))
-        };
-        self.text(
-            &truncate(&subtitle, 30),
-            x + 96.0,
-            y + 23.0,
-            9.0,
-            C_SUBTEXT0_CSS,
-            "normal",
-        );
-        // Direct line to the worker: open the composer targeted at the
-        // current prompt target (the dashboard resolves the routing).
-        self.pill_at(
-            x + panel_w - 132.0,
-            y + 13.0,
-            54.0,
-            23.0,
-            "steer",
-            C_BLUE_CSS,
-            false,
-        );
-        self.hit_zones.push(HitZone::new(
-            x + panel_w - 132.0,
-            y + 13.0,
-            54.0,
-            23.0,
-            HitAction::Composer { op: "open-send" },
-        ));
-
-        let mut row_y = y + 70.0;
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "source",
-            &format!(
-                "{} / {}",
-                nonempty(&agent.provider, "provider"),
-                nonempty(&agent.model, "model")
-            ),
-            C_BLUE_CSS,
-        );
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "phase",
-            &format!(
-                "{} / {}{}",
-                nonempty(&agent.phase, "idle"),
-                nonempty(&agent.status, "idle"),
-                if agent.autonomy.trim().is_empty() {
-                    String::new()
-                } else {
-                    format!(" / {} autonomy", agent.autonomy.trim())
-                }
-            ),
-            phase,
-        );
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "task",
-            &nonempty(&agent.task, "idle"),
-            C_TEAL_CSS,
-        );
-        if !agent.relationship_kind.trim().is_empty() {
-            let parent = agent
-                .parent_id
-                .as_deref()
-                .map(str::trim)
-                .filter(|p| !p.is_empty())
-                .map(|p| p.strip_prefix("session-").unwrap_or(p))
-                .map(|p| truncate(p, 12))
-                .unwrap_or_default();
-            let lineage = if parent.is_empty() {
-                agent.relationship_kind.trim().to_string()
-            } else {
-                format!("{} of {}", agent.relationship_kind.trim(), parent)
-            };
-            row_y = self.focus_row(x, row_y, panel_w, "lineage", &lineage, C_MAUVE_CSS);
-        }
-        if has_goal {
-            let status = agent.goal_status.trim();
-            let goal_color = goal_status_color_css(status);
-            let mut goal_text = if status.is_empty() {
-                agent.goal_objective.trim().to_string()
-            } else if agent.goal_objective.trim().is_empty() {
-                status.to_string()
-            } else {
-                format!("{}: {}", status, agent.goal_objective.trim())
-            };
-            if !agent.goal_tokens.trim().is_empty() {
-                goal_text.push_str(&format!(" ({} tok)", agent.goal_tokens.trim()));
-            }
-            row_y = self.focus_row(x, row_y, panel_w, "goal", &goal_text, goal_color);
-        }
-        let budget_pct = percent(agent.tokens, agent.token_cap);
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "tokens",
-            &format!(
-                "{} / {} ({})",
-                fmt_compact(agent.tokens),
-                fmt_compact(agent.token_cap),
-                pct_label(budget_pct)
-            ),
-            pressure_color(budget_pct),
-        );
-        self.meter(
-            x + 96.0,
-            row_y - 12.0,
-            panel_w - 116.0,
-            budget_pct,
-            pressure_color(budget_pct),
-        );
-        row_y += 6.0;
-        let mut usage = format!(
-            "p {} / c {} / cached {}",
-            fmt_compact(agent.prompt),
-            fmt_compact(agent.completion),
-            fmt_compact(agent.cached)
-        );
-        if agent.cost > 0.0 {
-            usage.push_str(&format!(" / ${:.2}", agent.cost));
-        }
-        if agent.turn_cap > 0 {
-            usage.push_str(&format!(" / turn {}/{}", agent.turns, agent.turn_cap));
-        } else if agent.turns > 0 {
-            usage.push_str(&format!(" / turn {}", agent.turns));
-        }
-        row_y = self.focus_row(x, row_y, panel_w, "usage", &usage, C_LAVENDER_CSS);
-        if !agent.vitals_git.trim().is_empty() {
-            let color = if agent.vitals_git_conflict {
-                C_RED_CSS
-            } else {
-                C_TEAL_CSS
-            };
-            row_y = self.focus_row(x, row_y, panel_w, "git", agent.vitals_git.trim(), color);
-        }
-        if agent.cache_hit_pct >= 0.0 || agent.cache_ttl_seconds > 0.0 {
-            // Same tiers as the dashboard chip (fragment 39): hit green
-            // ≥90 / yellow ≥50 / red, countdown red in its final minute,
-            // cold dimmed.
-            let mut text = String::new();
-            let mut color = C_SUBTEXT0_CSS;
-            if agent.cache_hit_pct >= 0.0 {
-                let hit = agent.cache_hit_pct.clamp(0.0, 100.0);
-                text.push_str(&format!("⚡{}%", hit.round() as u32));
-                color = if hit >= 90.0 {
-                    C_GREEN_CSS
-                } else if hit >= 50.0 {
-                    C_YELLOW_CSS
-                } else {
-                    C_PEACH_CSS
-                };
-            }
-            let now = epoch_seconds_now();
-            if agent.cache_ttl_seconds > 0.0 && agent.cache_last_activity_epoch > 0.0 && now > 0.0
-            {
-                let remaining =
-                    agent.cache_ttl_seconds - (now - agent.cache_last_activity_epoch);
-                if !text.is_empty() {
-                    text.push(' ');
-                }
-                if remaining > 0.0 {
-                    text.push_str(&format!("⏳{}", fmt_countdown(remaining)));
-                    if remaining <= 60.0 {
-                        color = C_RED_CSS;
-                    }
-                } else {
-                    text.push_str("✗ cold");
-                    color = C_SUBTEXT0_CSS;
-                }
-            }
-            if !text.is_empty() {
-                row_y = self.focus_row(x, row_y, panel_w, "cache", &text, color);
-            }
-        }
-        if !agent.vitals_limits.trim().is_empty() {
-            let color = match agent.vitals_limits_state.trim() {
-                "crit" => C_RED_CSS,
-                "warn" => C_YELLOW_CSS,
-                _ => C_SUBTEXT0_CSS,
-            };
-            row_y = self.focus_row(x, row_y, panel_w, "limits", agent.vitals_limits.trim(), color);
-        }
-        if !agent.worktree.trim().is_empty() {
-            row_y = self.focus_row(
-                x,
-                row_y,
-                panel_w,
-                "worktree",
-                agent.worktree.trim(),
-                C_MAUVE_CSS,
-            );
-        }
-
-        if is_session {
-            // Per-node action pills at session-window-kebab parity: the
-            // universal basics plus whatever thread-action ops the session
-            // advertises. Every pill dispatches through the dashboard's
-            // real session-action handler.
-            let sid = agent.session_id.trim().to_string();
-            let ops = &agent.thread_actions;
-            let py = row_y - 2.0;
-            let mut px = x + 96.0;
-            let mut pills: Vec<(&str, &str, &str)> =
-                vec![("log", C_BLUE_CSS, "station-log")];
-            if agent.recent {
-                // A closed session can be read or brought back — nothing
-                // else applies to it.
-                pills.push(("resume", C_GREEN_CSS, "resume"));
-            } else {
-                pills.push(("target", C_TEAL_CSS, "target"));
-                pills.push(("steer", C_LAVENDER_CSS, "steer"));
-                if agent.can_interrupt {
-                    pills.push(("stop", C_RED_CSS, "interrupt"));
-                }
-                if ops.iter().any(|op| op == "compact") {
-                    pills.push(("compact", C_MAUVE_CSS, "thread-compact"));
-                }
-                if ops.iter().any(|op| op == "fork") {
-                    pills.push(("fork", C_PEACH_CSS, "thread-fork"));
-                }
-            }
-            for (label, color, action) in pills {
-                let pw = label.chars().count() as f32 * 6.1 + 18.0;
-                if px + pw > x + panel_w - 16.0 {
-                    break;
-                }
-                self.pill_at(px, py, pw, 23.0, label, color, false);
-                self.hit_zones.push(HitZone::new(
-                    px,
-                    py,
-                    pw,
-                    23.0,
-                    HitAction::SessionAction {
-                        action: action.to_string(),
-                        id: sid.clone(),
-                    },
-                ));
-                px += pw + 8.0;
-            }
-            row_y += 32.0;
-        }
-
-        if approval {
-            row_y = self.focus_row(
-                x,
-                row_y,
-                panel_w,
-                "approval",
-                &format!(
-                    "{}{}",
-                    nonempty(&agent.approval_command, "approval required"),
-                    if agent.approval_category.trim().is_empty() {
-                        String::new()
-                    } else {
-                        format!(" ({})", agent.approval_category.trim())
-                    }
-                ),
-                C_YELLOW_CSS,
-            );
-            let host_id = agent.host_id.clone();
-            let approval_id = agent.approval_id.clone().unwrap_or_default();
-            let py = row_y - 6.0;
-            self.pill_at(x + 96.0, py, 78.0, 23.0, "approve", C_GREEN_CSS, false);
-            self.hit_zones.push(HitZone::new(
-                x + 96.0,
-                py,
-                78.0,
-                23.0,
-                HitAction::Approval {
-                    host_id: host_id.clone(),
-                    approval_id: approval_id.clone(),
-                    decision: "approve",
-                },
-            ));
-            self.pill_at(x + 182.0, py, 58.0, 23.0, "deny", C_RED_CSS, false);
-            self.hit_zones.push(HitZone::new(
-                x + 182.0,
-                py,
-                58.0,
-                23.0,
-                HitAction::Approval {
-                    host_id,
-                    approval_id,
-                    decision: "deny",
-                },
-            ));
-        }
-    }
-
-    /// View-settings panel for the system:view node: mood toggle pills plus
-    /// drag-aware fov/motion/AR/density sliders. Scrubs apply live in the
-    /// renderer; the released value is emitted as a `view_set` action that
-    /// the dashboard persists and re-applies through `set_visuals`.
-    fn draw_view_focus(&mut self, x: f32, panel_w: f32, activity_lane_y: f32) {
-        let panel_h = 74.0 + 30.0 + 4.0 * 26.0 + 12.0;
-        let y = (activity_lane_y - panel_h - 12.0).max(58.0);
-        self.focus_panel_frame(x, y, panel_w, panel_h, "View", C_LAVENDER_CSS);
-        self.text(
-            &format!("{} layout", self.layout.label()),
-            x + 96.0,
-            y + 23.0,
-            9.0,
-            C_SUBTEXT0_CSS,
-            "normal",
-        );
-
-        let mut row_y = y + 72.0;
-        self.text("mood", x + 16.0, row_y, 9.0, C_LAVENDER_CSS, "bold");
-        for (idx, mood) in [Mood::Cockpit, Mood::Calm].into_iter().enumerate() {
-            let px = x + 96.0 + idx as f32 * 86.0;
-            let label = mood.label();
-            self.pill_at(
-                px,
-                row_y - 16.0,
-                78.0,
-                23.0,
-                label,
-                if self.mood == mood {
-                    C_LAVENDER_CSS
-                } else {
-                    C_OVERLAY1_CSS
-                },
-                self.mood == mood,
-            );
-            self.hit_zones.push(HitZone::new(
-                px,
-                row_y - 16.0,
-                78.0,
-                23.0,
-                HitAction::ViewSet {
-                    key: "mood",
-                    value: label,
-                },
-            ));
-        }
-        row_y += 30.0;
-
-        let sliders = [
-            (
-                ViewSliderKey::Fov,
-                "fov",
-                format!("{}°", self.fov_deg.round() as i32),
-            ),
-            (
-                ViewSliderKey::Motion,
-                "motion",
-                format!("{:.1}x", self.motion),
-            ),
-            (
-                ViewSliderKey::Ar,
-                "ar tilt",
-                format!("{}%", (self.ar_strength * 100.0).round() as i32),
-            ),
-            (
-                ViewSliderKey::Density,
-                "density",
-                format!("{:.1}", self.density),
-            ),
-        ];
-        for (key, label, value_label) in sliders {
-            row_y = self.focus_slider(x, row_y, panel_w, key, label, &value_label);
-        }
-    }
-
-    /// One slider row: label, scrubbable track with fill + knob, value
-    /// readout. The hit zone is exactly the track rect (taller for touch),
-    /// which is also the geometry pointer x maps through.
-    fn focus_slider(
-        &mut self,
-        x: f32,
-        row_y: f32,
-        w: f32,
-        key: ViewSliderKey,
-        label: &str,
-        value_label: &str,
-    ) -> f32 {
-        self.text(label, x + 16.0, row_y, 9.0, C_LAVENDER_CSS, "bold");
-        let track_x = x + 96.0;
-        let track_w = w - 96.0 - 72.0;
-        let t = key.t_of(self.view_slider_value(key));
-        self.hud.set_fill("rgba(49,50,68,0.92)");
-        self.hud
-            .ctx
-            .fill_rect(track_x as f64, (row_y - 7.0) as f64, track_w as f64, 4.0);
-        self.hud.set_fill(C_LAVENDER_CSS);
-        self.hud.ctx.fill_rect(
-            track_x as f64,
-            (row_y - 7.0) as f64,
-            (track_w * t) as f64,
-            4.0,
-        );
-        self.hud.ctx.begin_path();
-        let _ = self.hud.ctx.arc(
-            (track_x + track_w * t) as f64,
-            (row_y - 5.0) as f64,
-            5.5,
-            0.0,
-            std::f64::consts::TAU,
-        );
-        self.hud.ctx.fill();
-        self.hud.set_stroke("rgba(17,17,27,0.9)");
-        self.hud.ctx.stroke();
-        self.text(value_label, x + w - 62.0, row_y, 9.0, C_TEXT_CSS, "normal");
-        self.hit_zones.push(HitZone::new(
-            track_x,
-            row_y - 16.0,
-            track_w,
-            22.0,
-            HitAction::ViewSlider { key },
-        ));
-        row_y + 26.0
-    }
-
-    /// Real detail panel for a selected host node: platform, link state,
-    /// load meters, and what is running / streaming on it.
-    fn draw_host_focus(
-        &mut self,
-        host: &crate::model::StationHost,
-        x: f32,
-        panel_w: f32,
-        activity_lane_y: f32,
-    ) {
-        let panel_h = 74.0 + 4.0 * 17.0 + 6.0;
-        let y = (activity_lane_y - panel_h - 12.0).max(58.0);
-        let color = if host.connected {
-            C_PEACH_CSS
-        } else {
-            C_RED_CSS
-        };
-        self.focus_panel_frame(x, y, panel_w, panel_h, &host.name, color);
-        self.text(
-            if host.connected {
-                "connected"
-            } else {
-                "offline"
-            },
-            x + 96.0,
-            y + 23.0,
-            9.0,
-            if host.connected {
-                C_GREEN_CSS
-            } else {
-                C_RED_CSS
-            },
-            "bold",
-        );
-        let mut row_y = y + 70.0;
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "platform",
-            &format!(
-                "{} / {}",
-                nonempty(&host.platform, "unknown"),
-                nonempty(&host.region, "local")
-            ),
-            C_BLUE_CSS,
-        );
-        let cpu_pct = (host.cpu / 100.0).clamp(0.0, 1.0);
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "cpu",
-            &pct_label(cpu_pct),
-            pressure_color(cpu_pct),
-        );
-        self.meter(
-            x + 156.0,
-            row_y - 12.0,
-            panel_w - 176.0,
-            cpu_pct,
-            pressure_color(cpu_pct),
-        );
-        let mem_pct = (host.mem / 100.0).clamp(0.0, 1.0);
-        row_y = self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "memory",
-            &pct_label(mem_pct),
-            pressure_color(mem_pct),
-        );
-        self.meter(
-            x + 156.0,
-            row_y - 12.0,
-            panel_w - 176.0,
-            mem_pct,
-            pressure_color(mem_pct),
-        );
-        let agents = self
-            .snapshot
-            .agents
-            .iter()
-            .filter(|a| a.host_id == host.id)
-            .count();
-        let waiting = self
-            .snapshot
-            .agents
-            .iter()
-            .filter(|a| a.host_id == host.id && a.needs_approval)
-            .count();
-        let streams = self
-            .display_sources
-            .values()
-            .filter(|s| s.host_id == host.id)
-            .count();
-        self.focus_row(
-            x,
-            row_y,
-            panel_w,
-            "running",
-            &format!(
-                "{agents} agent{} / {streams} stream{}{}",
-                if agents == 1 { "" } else { "s" },
-                if streams == 1 { "" } else { "s" },
-                if waiting > 0 {
-                    format!(" / {waiting} awaiting approval")
-                } else {
-                    String::new()
-                }
-            ),
-            C_TEAL_CSS,
-        );
-    }
-
-    pub(crate) fn station_focus_button(
-        &mut self,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-        target: &SystemTarget,
-    ) {
-        let SystemTarget {
-            id,
-            kicker,
-            title,
-            color,
-            ..
-        } = *target;
-        let value = &target.value;
-        let detail = &target.detail;
-        let selected = self.selected_id.as_deref() == Some(id);
-        let hovered = self
-            .hover_xy
-            .is_some_and(|(hx, hy)| hx >= x && hx <= x + w && hy >= y && hy <= y + h);
-        self.glass_panel(
-            x,
-            y,
-            w,
-            h,
-            8.0,
-            hex_color(color).unwrap_or(C_OVERLAY1),
-            if selected {
-                1.7
-            } else if hovered {
-                1.2
-            } else {
-                0.7
-            },
-            if selected { 1.1 } else { 0.85 },
-        );
-        self.hud.set_fill(color);
-        self.hud
-            .ctx
-            .fill_rect((x + 9.0) as f64, (y + 10.0) as f64, 4.0, (h - 20.0) as f64);
-        let max_chars = ((w - 34.0) / 6.2).max(12.0) as usize;
-        if h < 38.0 {
-            self.text(
-                &truncate(title, max_chars),
-                x + 20.0,
-                y + h * 0.5 + 4.0,
-                9.0,
-                color,
-                "bold",
-            );
-        } else if h < 58.0 {
-            self.text(title, x + 20.0, y + 18.0, 10.0, color, "bold");
-            self.text(
-                &truncate(value, max_chars),
-                x + 20.0,
-                y + 35.0,
-                9.5,
-                C_TEXT_CSS,
-                "normal",
-            );
-        } else if h < 72.0 {
-            if !kicker.is_empty() {
-                self.text(kicker, x + 20.0, y + 15.0, 7.5, C_OVERLAY1_CSS, "bold");
-            }
-            self.text(
-                title,
-                x + 20.0,
-                y + if kicker.is_empty() { 21.0 } else { 29.0 },
-                10.5,
-                color,
-                "bold",
-            );
-            self.text(
-                &truncate(value, max_chars),
-                x + 20.0,
-                y + if detail.is_empty() {
-                    h - 13.0
-                } else {
-                    h - 25.0
-                },
-                9.5,
-                C_TEXT_CSS,
-                "normal",
-            );
-            if !detail.is_empty() {
-                self.text(
-                    &truncate(detail, max_chars),
-                    x + 20.0,
-                    y + h - 11.0,
-                    8.0,
-                    C_SUBTEXT0_CSS,
-                    "normal",
-                );
-            }
-        } else {
-            if !kicker.is_empty() {
-                self.text(kicker, x + 20.0, y + 16.0, 8.0, C_OVERLAY1_CSS, "bold");
-            }
-            self.text(
-                title,
-                x + 20.0,
-                y + if kicker.is_empty() { 24.0 } else { 34.0 },
-                11.0,
-                color,
-                "bold",
-            );
-            self.text(
-                &truncate(value, max_chars),
-                x + 20.0,
-                y + h - if detail.is_empty() { 15.0 } else { 29.0 },
-                10.0,
-                C_TEXT_CSS,
-                "normal",
-            );
-            if !detail.is_empty() {
-                self.text(
-                    &truncate(detail, max_chars),
-                    x + 20.0,
-                    y + h - 12.0,
-                    8.5,
-                    C_SUBTEXT0_CSS,
-                    "normal",
-                );
-            }
-        }
-        self.hit_zones
-            .push(HitZone::new(x, y, w, h, HitAction::Select(id.to_string())));
-    }
-
-    pub(crate) fn station_target_label(&self) -> String {
-        let controls = &self.snapshot.controls;
-        nonempty(
-            &controls.session_label,
-            &nonempty(
-                &controls.session_selection,
-                &nonempty(&controls.command, "No active command target"),
-            ),
-        )
-    }
-
-    pub(crate) fn station_primary_actions(&self) -> Vec<LaneAction> {
-        let controls = &self.snapshot.controls;
-        // send/new session open the in-canvas composer (send + launch
-        // modes) — they used to focus inputs on hidden dashboard tabs.
-        let mut actions = vec![
-            LaneAction::composer(
-                if controls.prompt_mode == "steer" {
-                    "steer"
-                } else {
-                    "send"
-                },
-                "open-send",
-                72.0,
-                C_BLUE_CSS,
-            ),
-            LaneAction::composer("new session", "open-launch", 112.0, C_TEAL_CSS),
-        ];
-        if controls.session_can_focus {
-            actions.push(LaneAction::activity("focus", "target", 72.0, C_PEACH_CSS));
-        }
-        if controls.session_can_interrupt {
-            actions.push(LaneAction::activity("stop", "stop", 60.0, C_RED_CSS));
-        }
-        if controls.shared_view_can_take_input {
-            actions.push(LaneAction::controls(
-                "take input",
-                "shared-view-take-input",
-                102.0,
-                C_GREEN_CSS,
-            ));
-        }
-        actions.extend([
-            LaneAction::select("context", "system:context", 82.0, C_BLUE_CSS),
-            LaneAction::select("managed", "system:managed", 88.0, C_MAUVE_CSS),
-            LaneAction::select("sessions", "system:sessions", 90.0, C_TEAL_CSS),
-            LaneAction::select("controls", "system:controls", 88.0, C_MAUVE_CSS),
-        ]);
-        actions
-    }
-
-    pub(crate) fn compute_system_targets(&self) -> Vec<SystemTarget> {
-        let latest_event = self.snapshot.events.last();
-        let ctx_pct = percent(
-            self.snapshot.context.tokens,
-            self.snapshot.context.effective_window,
-        );
-        let managed_pct = percent(
-            self.snapshot.managed.used_tokens,
-            self.snapshot.managed.effective_window,
-        );
-        let changes = &self.snapshot.changes;
-        let controls = &self.snapshot.controls;
-        let peer_count = self.snapshot.hosts.len().saturating_sub(1);
-        vec![
-            SystemTarget {
-                id: "system:activity",
-                kicker: "signal",
-                title: "Activity",
-                value: format!("{} retained", activity_retained_count(&self.snapshot)),
-                detail: latest_event
-                    .map(|event| truncate(&format!("{} {}", event.level, event.msg), 30))
-                    .unwrap_or_else(|| "waiting for events".to_string()),
-                color: latest_event
-                    .map(|event| level_color_css(&event.level))
-                    .unwrap_or(C_TEAL_CSS),
-            },
-            SystemTarget {
-                id: "system:context",
-                kicker: "memory",
-                title: "Context",
-                value: if self.snapshot.context.available {
-                    format!(
-                        "{} / {} items",
-                        pct_label(ctx_pct),
-                        self.snapshot.context.item_count
-                    )
-                } else {
-                    "waiting".to_string()
-                },
-                detail: truncate(
-                    &format!(
-                        "{} {}",
-                        nonempty(&self.snapshot.context.source, "snapshot"),
-                        nonempty(&self.snapshot.context.turn, "")
-                    ),
-                    30,
-                ),
-                color: pressure_color(ctx_pct),
-            },
-            SystemTarget {
-                id: "system:managed",
-                kicker: "lineage",
-                title: "Managed",
-                value: format!(
-                    "{} / {}",
-                    nonempty(&self.snapshot.managed.mode, "managed"),
-                    nonempty(&self.snapshot.managed.status, "unknown")
-                ),
-                detail: format!(
-                    "{} records / {} anchors",
-                    self.snapshot.managed.records, self.snapshot.managed.anchors
-                ),
-                color: pressure_color(managed_pct),
-            },
-            SystemTarget {
-                id: "system:controls",
-                kicker: "operator",
-                title: "Controls",
-                value: truncate(
-                    &format!(
-                        "{} / {}",
-                        nonempty(&controls.backend, "agent"),
-                        nonempty(&controls.sandbox, "sandbox")
-                    ),
-                    32,
-                ),
-                detail: truncate(
-                    &format!(
-                        "{} / managed {}",
-                        nonempty(&controls.approval_policy, "approval"),
-                        nonempty(&controls.managed_context, "unknown")
-                    ),
-                    34,
-                ),
-                color: C_MAUVE_CSS,
-            },
-            SystemTarget {
-                id: "system:sessions",
-                kicker: "work",
-                title: "Sessions",
-                value: format!(
-                    "{} total / {} active",
-                    self.snapshot.sessions.total, self.snapshot.sessions.active
-                ),
-                detail: truncate(
-                    &nonempty(&self.snapshot.sessions.latest_task, "launch history"),
-                    32,
-                ),
-                color: if self.snapshot.sessions.active > 0 {
-                    C_TEAL_CSS
-                } else {
-                    C_BLUE_CSS
-                },
-            },
-            SystemTarget {
-                id: "system:peers",
-                kicker: "display",
-                title: "Peers",
-                value: format!(
-                    "{peer_count} peers / {} streams",
-                    self.display_sources.len()
-                ),
-                detail: truncate(
-                    &format!(
-                        "{} / {}",
-                        nonempty(&controls.display_access, "display"),
-                        nonempty(&controls.cu_backend, "computer use")
-                    ),
-                    34,
-                ),
-                color: C_PEACH_CSS,
-            },
-            SystemTarget {
-                id: "system:changes",
-                kicker: "tree",
-                title: "Changes",
-                value: if changes.count > 0 {
-                    format!(
-                        "{} files / +{} -{}",
-                        changes.count, changes.total_added, changes.total_removed
-                    )
-                } else {
-                    nonempty(&changes.status, "clean")
-                },
-                detail: truncate(&nonempty(&changes.latest_path, "working tree clean"), 34),
-                color: if changes.count > 0 || changes.status == "mismatch" {
-                    C_YELLOW_CSS
-                } else {
-                    C_GREEN_CSS
-                },
-            },
-            SystemTarget {
-                id: "system:worktrees",
-                kicker: "project",
-                title: "Worktrees",
-                value: format!(
-                    "{} scanned / {} active",
-                    self.snapshot.sessions.worktrees, self.snapshot.sessions.worktree_active
-                ),
-                detail: format!(
-                    "{} dirty / {} unmerged",
-                    self.snapshot.sessions.worktree_dirty, self.snapshot.sessions.worktree_unmerged
-                ),
-                color: if self.snapshot.sessions.worktree_dirty > 0
-                    || self.snapshot.sessions.worktree_unmerged > 0
-                {
-                    C_YELLOW_CSS
-                } else {
-                    C_BLUE_CSS
-                },
-            },
-            SystemTarget {
-                id: "system:view",
-                kicker: "scene",
-                title: "View",
-                value: format!("{} / {}", self.layout.label(), self.mood.label()),
-                detail: format!(
-                    "{} fov / {:.1} density",
-                    self.fov_deg.round() as i32,
-                    self.density
-                ),
-                color: C_LAVENDER_CSS,
-            },
-        ]
-    }
-
-    pub(crate) fn draw_corners(&self, w: f32, h: f32) {
-        let a = self.mood.glass();
-        self.hud
-            .set_stroke(&css_rgba(C_LAVENDER.with_alpha(0.34 * a).into()));
-        let len = 26.0;
-        for (x, y, sx, sy) in [
-            (11.0, 50.0, 1.0, 1.0),
-            (w - 11.0, 50.0, -1.0, 1.0),
-            (11.0, h - 11.0, 1.0, -1.0),
-            (w - 11.0, h - 11.0, -1.0, -1.0),
-        ] {
-            self.line(x, y, x + sx * len, y);
-            self.line(x, y, x, y + sy * len);
-        }
-    }
-
-    pub(crate) fn draw_compass(&self, w: f32, h: f32) {
-        let cx = w - 71.0;
-        let cy = h - 33.0;
-        // Small glass disc so the dial reads over any scene behind it.
-        self.hud.ctx.begin_path();
-        let _ = self
-            .hud
-            .ctx
-            .arc(cx as f64, cy as f64, 18.0, 0.0, std::f64::consts::TAU);
-        self.hud.set_fill("rgba(13,14,24,0.55)");
-        self.hud.ctx.fill();
-        self.hud.set_stroke(&css_rgba(
-            C_LAVENDER.with_alpha(0.40 * self.mood.glass()).into(),
-        ));
-        self.hud.ctx.stroke();
-        let angle = -self.yaw as f64;
-        self.hud.set_stroke(C_BLUE_CSS);
-        self.hud.ctx.begin_path();
-        self.hud.ctx.move_to(cx as f64, cy as f64);
-        self.hud.ctx.line_to(
-            cx as f64 + angle.sin() * 14.0,
-            cy as f64 - angle.cos() * 14.0,
-        );
-        self.hud.ctx.stroke();
-        self.text("N", cx + 27.0, cy + 4.0, 10.0, C_OVERLAY1_CSS, "bold");
-    }
-
-    pub(crate) fn meter(&self, x: f32, y: f32, w: f32, pct: f32, color: &str) {
-        let pct = pct.clamp(0.0, 1.0);
-        self.hud.set_fill("rgba(49,50,68,0.92)");
-        self.hud
-            .ctx
-            .fill_rect(x as f64, (y - 6.0) as f64, w as f64, 5.0);
-        self.hud.set_fill(color);
-        self.hud
-            .ctx
-            .fill_rect(x as f64, (y - 6.0) as f64, (w * pct) as f64, 5.0);
-        self.hud.set_stroke("rgba(127,132,156,0.5)");
-        self.hud
-            .ctx
-            .stroke_rect(x as f64, (y - 6.0) as f64, w as f64, 5.0);
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn pill_button(
-        &mut self,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-        label: &str,
-        active: bool,
-        action: HitAction,
-    ) {
-        self.pill_at(
-            x,
-            y,
-            w,
-            h,
-            label,
-            if active { C_BLUE_CSS } else { C_OVERLAY1_CSS },
-            active,
-        );
-        self.hit_zones.push(HitZone::new(x, y, w, h, action));
-    }
-
-    /// Capsule pill with the glass treatment. `active` (selected) and
-    /// hovered pills are lit from within: an accent gradient swelling from
-    /// the capsule's middle plus a brighter luminous border.
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn pill_at(
-        &self,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-        label: &str,
-        color: &str,
-        active: bool,
-    ) {
-        let ctx = &self.hud.ctx;
-        let a = self.mood.glass();
-        let accent = hex_color(color).unwrap_or(C_OVERLAY1);
-        let hovered = self
-            .hover_xy
-            .is_some_and(|(hx, hy)| hx >= x && hx <= x + w && hy >= y && hy <= y + h);
-        let r = (h * 0.5).min(11.0);
-        // Dark translucent capsule base.
-        self.rounded_path(x, y, w, h, r);
-        let base = ctx.create_linear_gradient(x as f64, y as f64, x as f64, (y + h) as f64);
-        let _ = base.add_color_stop(
-            0.0,
-            &css_rgba(Color::rgb(42, 44, 66).with_alpha(0.52).into()),
-        );
-        let _ = base.add_color_stop(
-            1.0,
-            &css_rgba(Color::rgb(13, 14, 24).with_alpha(0.68).into()),
-        );
-        ctx.set_fill_style_canvas_gradient(&base);
-        self.hud.note_fill_unknown();
-        ctx.fill();
-        if active || hovered {
-            let lit = (if active { 0.30 } else { 0.20 }) * a;
-            let inner = ctx.create_linear_gradient(x as f64, y as f64, x as f64, (y + h) as f64);
-            let _ = inner.add_color_stop(0.0, &css_rgba(accent.with_alpha(lit * 0.35).into()));
-            let _ = inner.add_color_stop(0.5, &css_rgba(accent.with_alpha(lit).into()));
-            let _ = inner.add_color_stop(1.0, &css_rgba(accent.with_alpha(lit * 0.45).into()));
-            ctx.set_fill_style_canvas_gradient(&inner);
-            ctx.fill();
-        }
-        // Gentle top highlight, then the luminous border.
-        self.hud.set_stroke(&css_rgba([0.93, 0.95, 1.0, 0.07 * a]));
-        self.line(x + r, y + 1.0, x + w - r, y + 1.0);
-        let border = if active {
-            0.85
-        } else if hovered {
-            0.62
-        } else {
-            0.38
-        } * a;
-        self.rounded_path(x, y, w, h, r);
-        self.hud
-            .set_stroke(&css_rgba(accent.with_alpha(border).into()));
-        ctx.stroke();
-        self.text(label, x + 8.0, y + h * 0.65, 10.0, color, "bold");
-    }
-
-    /// Trace a rounded-rect path on the HUD context (no fill/stroke).
-    pub(crate) fn rounded_path(&self, x: f32, y: f32, w: f32, h: f32, r: f32) {
-        let ctx = &self.hud.ctx;
-        let r = r.min(w * 0.5).min(h * 0.5).max(0.0);
-        ctx.begin_path();
-        ctx.move_to((x + r) as f64, y as f64);
-        ctx.line_to((x + w - r) as f64, y as f64);
-        ctx.quadratic_curve_to((x + w) as f64, y as f64, (x + w) as f64, (y + r) as f64);
-        ctx.line_to((x + w) as f64, (y + h - r) as f64);
-        ctx.quadratic_curve_to(
-            (x + w) as f64,
-            (y + h) as f64,
-            (x + w - r) as f64,
-            (y + h) as f64,
-        );
-        ctx.line_to((x + r) as f64, (y + h) as f64);
-        ctx.quadratic_curve_to(x as f64, (y + h) as f64, x as f64, (y + h - r) as f64);
-        ctx.line_to(x as f64, (y + r) as f64);
-        ctx.quadratic_curve_to(x as f64, y as f64, (x + r) as f64, y as f64);
-        ctx.close_path();
-    }
-
-    /// Frosted-glass panel, canvas-native: a soft outer shadow, layered
-    /// translucent body gradient, a top-edge specular sheen, a faint inner
-    /// highlight, and a 1px luminous border with corner glow. Everything is
-    /// plain gradient/alpha layering — no `ctx.filter` blur, which would be
-    /// far too slow to repaint per frame.
-    ///
-    /// `emphasis` scales the accent (border/corner) luminosity — ~1.0 for
-    /// resting panels, higher for selected/featured ones. `tint` scales the
-    /// body opacity — 1.0 for solid panels, low values for see-through
-    /// surfaces that must not hide the 3D scene behind them. The calm mood
-    /// additionally dims all accents via [`Mood::glass`].
-    #[allow(clippy::too_many_arguments)]
-    pub(crate) fn glass_panel(
-        &self,
-        x: f32,
-        y: f32,
-        w: f32,
-        h: f32,
-        r: f32,
-        accent: Color,
-        emphasis: f32,
-        tint: f32,
-    ) {
-        let ctx = &self.hud.ctx;
-        let a = self.mood.glass();
-        // Soft outer shadow: one slightly enlarged, downward-biased dark
-        // fill fakes a blurred drop shadow.
-        self.rounded_path(x - 2.0, y - 1.0, w + 4.0, h + 5.0, r + 3.0);
-        self.hud.set_fill("rgba(2,3,9,0.30)");
-        ctx.fill();
-        // Body: deep dark vertical gradient (lighter up top, denser below).
-        self.rounded_path(x, y, w, h, r);
-        let body = ctx.create_linear_gradient(x as f64, y as f64, x as f64, (y + h) as f64);
-        let _ = body.add_color_stop(
-            0.0,
-            &css_rgba(Color::rgb(38, 40, 60).with_alpha(0.62 * tint).into()),
-        );
-        let _ = body.add_color_stop(
-            0.45,
-            &css_rgba(Color::rgb(21, 22, 34).with_alpha(0.74 * tint).into()),
-        );
-        let _ = body.add_color_stop(
-            1.0,
-            &css_rgba(Color::rgb(12, 12, 20).with_alpha(0.85 * tint).into()),
-        );
-        ctx.set_fill_style_canvas_gradient(&body);
-        self.hud.note_fill_unknown();
-        ctx.fill();
-        // Top-edge specular sheen; the body path is still current.
-        let sheen_h = (h * 0.42).clamp(8.0, 30.0);
-        let sheen = ctx.create_linear_gradient(x as f64, y as f64, x as f64, (y + sheen_h) as f64);
-        let _ = sheen.add_color_stop(0.0, &css_rgba([0.92, 0.95, 1.0, 0.10 * a]));
-        let _ = sheen.add_color_stop(1.0, "rgba(235,242,255,0)");
-        ctx.set_fill_style_canvas_gradient(&sheen);
-        ctx.fill();
-        // Gentle inner highlight stroke, inset 1px.
-        self.rounded_path(
-            x + 1.0,
-            y + 1.0,
-            (w - 2.0).max(1.0),
-            (h - 2.0).max(1.0),
-            (r - 1.0).max(1.5),
-        );
-        self.hud.set_stroke(&css_rgba([0.93, 0.95, 1.0, 0.05 * a]));
-        ctx.stroke();
-        // 1px luminous border.
-        let border = ((0.26 + 0.26 * emphasis) * a).min(0.92);
-        self.rounded_path(x, y, w, h, r);
-        self.hud
-            .set_stroke(&css_rgba(accent.with_alpha(border).into()));
-        ctx.stroke();
-        // Corner glow: brighter quarter-arcs hugging each rounded corner.
-        let glow = (0.55 * emphasis * a).min(0.95);
-        self.hud
-            .set_stroke(&css_rgba(accent.with_alpha(glow).into()));
-        let cr = r.max(2.0).min(w * 0.5).min(h * 0.5) as f64;
-        let half_pi = std::f64::consts::FRAC_PI_2;
-        for (cx, cy, start) in [
-            (x + cr as f32, y + cr as f32, std::f64::consts::PI),
-            (x + w - cr as f32, y + cr as f32, 1.5 * std::f64::consts::PI),
-            (x + w - cr as f32, y + h - cr as f32, 0.0),
-            (x + cr as f32, y + h - cr as f32, half_pi),
-        ] {
-            ctx.begin_path();
-            let _ = ctx.arc(cx as f64, cy as f64, cr, start, start + half_pi);
-            ctx.stroke();
-        }
-    }
-
-    pub(crate) fn text(&self, text: &str, x: f32, y: f32, px: f32, color: &str, weight: &str) {
-        self.hud.set_fill(color);
-        self.hud.set_font(px, weight == "bold");
-        let _ = self.hud.ctx.fill_text(text, x as f64, y as f64);
-    }
-
-    pub(crate) fn line(&self, x1: f32, y1: f32, x2: f32, y2: f32) {
-        self.hud.ctx.begin_path();
-        self.hud.ctx.move_to(x1 as f64, y1 as f64);
-        self.hud.ctx.line_to(x2 as f64, y2 as f64);
-        self.hud.ctx.stroke();
-    }
-
-    pub(crate) fn css_width(&self) -> f32 {
-        self.width as f32 / self.dpr as f32
-    }
-
-    pub(crate) fn css_height(&self) -> f32 {
-        self.height as f32 / self.dpr as f32
     }
 }
 
@@ -3454,66 +526,6 @@ mod tests {
                 "density {density}: row {last_row} vs lane {height}"
             );
         }
-    }
-
-    #[test]
-    fn wrap_line_wraps_words_and_hard_breaks_long_tokens() {
-        assert_eq!(wrap_line("short line", 20), vec!["short line"]);
-        let wrapped = wrap_line("alpha beta gamma delta epsilon", 11);
-        assert!(
-            wrapped.iter().all(|l| l.chars().count() <= 11),
-            "{wrapped:?}"
-        );
-        assert_eq!(wrapped.join(" "), "alpha beta gamma delta epsilon");
-        // Overlong tokens hard-break instead of overflowing the panel.
-        let token = "a".repeat(30);
-        let broken = wrap_line(&token, 10);
-        assert_eq!(broken.len(), 3);
-        assert!(broken.iter().all(|l| l.chars().count() <= 10));
-        // Empty input still yields one (empty) line.
-        assert_eq!(wrap_line("", 10), vec![""]);
-    }
-
-    #[test]
-    fn transcript_layout_assigns_monotonic_offsets_and_marks_first_lines() {
-        let transcript = crate::model::StationTranscript {
-            session_id: "s1".into(),
-            rows: vec![
-                crate::model::StationTranscriptRow {
-                    kind: "user".into(),
-                    ts: "12:00".into(),
-                    text: "do the thing".into(),
-                },
-                crate::model::StationTranscriptRow {
-                    kind: "model".into(),
-                    ts: "12:01".into(),
-                    text: "first paragraph that is long enough to wrap across lines\nsecond line"
-                        .into(),
-                },
-            ],
-            ..Default::default()
-        };
-        let layout = layout_transcript(&transcript, 24);
-        assert!(layout.lines.len() >= 4);
-        assert!(layout.lines[0].first);
-        assert_eq!(layout.lines[0].kind, "user");
-        // Exactly one `first` line per row.
-        assert_eq!(layout.lines.iter().filter(|l| l.first).count(), 2);
-        // Offsets strictly increase and content height covers them all.
-        for pair in layout.lines.windows(2) {
-            assert!(pair[1].y > pair[0].y);
-        }
-        assert!(layout.content_h >= layout.lines.last().unwrap().y);
-        // Signature changes when content grows.
-        let sig_a = transcript_signature(&transcript, 24);
-        let mut grown = transcript.clone();
-        grown.rows.push(crate::model::StationTranscriptRow {
-            kind: "agent".into(),
-            ts: String::new(),
-            text: "tail".into(),
-        });
-        assert_ne!(sig_a, transcript_signature(&grown, 24));
-        assert_ne!(sig_a, transcript_signature(&transcript, 30));
     }
 
     #[test]
